fix: resolve all remaining MyPy errors - achieve 0 errors in 408 files (batch 3)
- Add mamba_ssm, face_recognition, facenet_pytorch, lime, shap, tomli,
  ava_guardian, aiohttp to ignore_missing_imports in pyproject.toml
- Fix no-any-return errors with explicit casts in core, biometric,
  harmonics, explainability, and models modules
- Fix variable shadowing causing type conflicts in biometric/__init__.py
  and core/gosnn_integration.py
- Fix union-attr errors with isinstance narrowing in explainability modules
- Fix var-annotated, no-untyped-def, type-arg, assignment errors
- Remove now-redundant type: ignore[import-not-found] comments
- MyPy now reports: "Success: no issues found in 408 source files"

+++ src/omni_mercury_engine/explainability/shap.py
@@ -247,7 +247,7 @@ class ExactShapExplainer(ShapExplainer):
                 combined[j] = x[j]
             predictions[i] = self._predict(combined.reshape(1, -1))[0]
 
-        return np.mean(predictions)
+        return float(np.mean(predictions))
 
     def explain_global(self, X: np.ndarray) -> GlobalExplanation:
         """Compute global SHAP explanation."""
@@ -389,7 +389,7 @@ class KernelShapExplainer(ShapExplainer):
             pred = self._predict(combined.reshape(1, -1))[0]
             predictions.append(pred)
 
-        return np.mean(predictions)
+        return float(np.mean(predictions))
 
     def explain_global(self, X: np.ndarray) -> GlobalExplanation:
         """Compute global Kernel SHAP explanation."""
@@ -739,4 +739,4 @@ def create_shap_explainer(
     if explainer_type not in explainer_map:
         raise ValueError(f"Unknown explainer type: {explainer_type}")
 
-    return explainer_map[explainer_type]()
+    return explainer_map[explainer_type]()  # type: ignore[no-untyped-call]
